requirements.txt
pandas
numpy
jupyter
scikit-learn
python-box
matplotlib
seaborn
fastapi
pydantic
uvicorn
textblob
nltk
-e .